fixed warnings.

+++ adapter/test/posix/posix_adapter_mpi_test.cpp
@@ -9,7 +9,7 @@ namespace hermes::adapter::posix::test {
 struct Arguments {
     std::string filename = "test.dat";
     std::string directory = "/tmp";
-    long request_size = 16384;
+    size_t request_size = 16384;
 };
 struct Info {
     int rank = 0;
@@ -20,7 +20,7 @@ struct Info {
     std::string existing_file;
     std::string shared_new_file;
     std::string shared_existing_file;
-    long num_iterations = 1024;
+    size_t num_iterations = 1024;
     unsigned int offset_seed = 1;
     unsigned int rs_seed = 1;
     unsigned int temporal_interval_seed = 1;
