+++ adapter/test/stdio/stdio_adapter_mpi_test.cpp
@@ -10,7 +10,7 @@ namespace hermes::adapter::stdio::test {
 struct Arguments {
     std::string filename = "test.dat";
     std::string directory = "/tmp";
-    long request_size = 16384;
+    size_t request_size = 16384;
 };
 struct Info {
     int rank = 0;
@@ -19,7 +19,7 @@ struct Info {
     std::string read_data;
     std::string new_file;
     std::string existing_file;
-    long num_iterations = 1024;
+    size_t num_iterations = 1024;
     unsigned int offset_seed = 1;
     unsigned int rs_seed = 1;
     unsigned int temporal_interval_seed = 1;
